[Refactor] Move consign app to Scenes folder
