+++ src/lib/Containers/Inquiry.tsx
@@ -3,12 +3,12 @@ import { InquiryQuery } from "__generated__/InquiryQuery.graphql"
 import colors from "lib/data/colors"
 import fonts from "lib/data/fonts"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
+import { BottomAlignedButton } from "lib/Scenes/Consignments/Components/BottomAlignedButton"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 import React from "react"
 import { Dimensions, NativeModules } from "react-native"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
 import styled from "styled-components/native"
-import { BottomAlignedButton } from "../Components/Consignments/Components/BottomAlignedButton"
 import ArtworkPreview from "../Components/Inbox/Conversations/Preview/ArtworkPreview"
 import { MetadataText, SmallHeadline } from "../Components/Inbox/Typography"
 import ARSwitchBoard from "../NativeModules/SwitchBoard"
