+++ src/lib/Scenes/Consignments/Submission/Gemini/getGeminiCredentialsForEnvironment.ts
