+++ src/lib/Scenes/Consignments/Submission/uploadPhotoToGemini.ts
